requirements.txt
fastapi
streamlit
euriai
pytesseract
python-dotenv